+++ bookmark.lua
@@ -1,4 +1,4 @@
-VERSION = "2.3.11"
+VERSION = "2.4.0"
 
 local micro    = import("micro")
 local buffer   = import("micro/buffer")